+++ pkgs/firehose/lib/src/utils.dart
@@ -59,9 +59,9 @@ class CommandResult {
 class Tag {
   /// RegExp matching a version tag at the start of a line.
   ///
-  /// A version tag is an optional starting seqeuence of non-whitespace, which
-  /// is the package name, followed by a `v` and a simplified SemVer version
-  /// number.
+  /// A version tag is an optional starting sequence of non-whitespace, which
+  /// is the package name, followed by [prefix] (defaulting to `'v'`) and a
+  /// simplified SemVer version number.
   ///
   /// The version number accepted is:
   ///
@@ -69,37 +69,40 @@ class Tag {
   ///
   /// and if followed by a `+` or `-`, then it includes the remaining
   /// non-whitespace characters.
-  static final RegExp packageVersionTag =
-      RegExp(r'^(?:(\S+)-)?v(\d+\.\d+\.\d+(?:[+\-]\S*)?)');
+  static RegExp packageVersionTag({String prefix = 'v'}) => RegExp(
+      '^(?:(\\S+)-)?${RegExp.escape(prefix)}(\\d+\\.\\d+\\.\\d+(?:[+\\-]\\S*)?)');
 
   /// A package version tag.
   ///
   /// Is expected to have the format:
   ///
-  /// > (package-name)? 'v' SemVer-version
+  /// > (package-name)? prefix SemVer-version
   ///
   /// If not, the tag is not [valid], and the [package] and [version] will both
   /// be `null`.
   final String tag;
+  final String prefix;
 
-  Tag(this.tag);
+  Tag(this.tag, {this.prefix = 'v'});
+
+  late final Match? _match = packageVersionTag(prefix: prefix).firstMatch(tag);
 
   bool get valid => version != null;
 
-  /// The package name before the `v` in the version [tag], if any.
+  /// The package name before the prefix in the version [tag], if any.
   ///
-  /// Is `null` if there is no package name before the `v`,
+  /// Is `null` if there is no package name before the prefix,
   /// or if the tag is not [valid].
-  String? get package => packageVersionTag.firstMatch(tag)?[1];
+  String? get package => _match?[1];
 
   /// The SemVer version string of the version [tag], if any.
   ///
-  /// This is the part after the `v` of the [tag] string,
+  /// This is the part after the prefix of the [tag] string,
   /// of the form, which is a major/minor/patch version string
   /// optionally followed by a `+` and more characters.
   ///
   /// Is `null` if the tag is not [valid].
-  String? get version => packageVersionTag.firstMatch(tag)?[2];
+  String? get version => _match?[2];
 
   @override
   String toString() => tag;